llama3.2-11b-instruct/requirements.txt
accelerate==1.2.1
bentoml>=1.3.16
fastapi==0.115.6
openai==1.57.4
pillow==10.4.0
vllm==0.6.6.post1; sys_platform == "linux"
